html/semantics/scripting-1/the-script-element/script-defer.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: HTMLScriptElement - defer</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script>

let ran_deferred_script = true;
let t = async_test("the defer script run later");

</script>

<script type="text/javascript" src="defer.js" defer></script>

<script>

t.step(() => {
  assert_true(ran_deferred_script);
  ran_deferred_script = false;
});

</script>
